Implemented Cross-User Data Access

+++ src/tasks/tasks.controller.ts
@@ -3,6 +3,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   HttpStatus,
@@ -11,6 +12,7 @@ import {
   Patch,
   Post,
   Query,
+  Request,
 } from '@nestjs/common';
 import { TasksService } from './tasks.service';
 import { CreateTaskDto } from './create-task.dto';
@@ -22,6 +24,8 @@ import { CreateTaskLabelDto } from './create-task-label.dto';
 import { FindTaskParams } from './find-task.params';
 import { PaginationParams } from './../common/pagination.params';
 import { PaginationResponse } from './../common/pagination.response';
+import { CurrentUserId } from '../users/decorators/current-user-id.decorator';
+// import { AuthRequest } from './../users/auth.request';
 
 @Controller('tasks')
 export class TasksController {
@@ -30,8 +34,13 @@ export class TasksController {
   public async findAll(
     @Query() filters: FindTaskParams,
     @Query() pagination: PaginationParams,
+    @CurrentUserId() userId: string,
   ): Promise<PaginationResponse<Task>> {
-    const [items, total] = await this.tasksService.findAll(filters, pagination);
+    const [items, total] = await this.tasksService.findAll(
+      filters,
+      pagination,
+      userId,
+    );
     return {
       data: items,
       meta: {
@@ -43,16 +52,26 @@ export class TasksController {
   }
 
   @Get('/:id')
-  public async findOne(@Param() params: FindOneParams): Promise<Task> {
-    return await this.findOneOrFail(params.id);
+  public async findOne(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
+    return task;
   }
 
   @Post()
-  public async createTask(@Body() createTaskDto: CreateTaskDto): Promise<Task> {
-    return await this.tasksService.createTask(createTaskDto);
+  public async createTask(
+    @Body() createTaskDto: CreateTaskDto,
+    // @Request() request: AuthRequest,
+    @CurrentUserId() userId: string,
+  ): Promise<Task> {
+    return await this.tasksService.createTask(
+      Object.assign({}, createTaskDto, { userId }),
+    );
   }
 
-  // @Patch('/:id/status')
   // public async updateTaskStatus(
   //   @Param() params: FindOneParams,
   //   @Body() body: UpdateTaskStatusDto,
@@ -73,9 +92,10 @@ export class TasksController {
   public async updateTask(
     @Param() params: FindOneParams,
     @Body() updateTaskDto: UpdateTaskDto,
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
-    console.log('updateTask');
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
     try {
       return await this.tasksService.updateTask(task, updateTaskDto);
     } catch (error) {
@@ -88,8 +108,12 @@ export class TasksController {
 
   @Delete('/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  public async deleteTask(@Param() params: FindOneParams): Promise<void> {
+  public async deleteTask(
+    @Param() params: FindOneParams,
+    @CurrentUserId() userId: string,
+  ): Promise<void> {
     const task = await this.findOneOrFail(params.id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.deleteTask(task);
   }
 
@@ -98,8 +122,10 @@ export class TasksController {
   public async removeLabels(
     @Param() { id }: FindOneParams,
     @Body() labelNames: string[],
+    @CurrentUserId() userId: string,
   ): Promise<void> {
     const task = await this.findOneOrFail(id);
+    this.checkTaskOwnership(task, userId);
     await this.tasksService.removeLabels(task, labelNames);
   }
 
@@ -110,8 +136,10 @@ export class TasksController {
   async addLabels(
     @Param() { id }: FindOneParams,
     @Body() labels: CreateTaskLabelDto[],
+    @CurrentUserId() userId: string,
   ): Promise<Task> {
     const task = await this.findOneOrFail(id);
+    this.checkTaskOwnership(task, userId);
     return await this.tasksService.addLabels(task, labels);
   }
 
@@ -121,6 +149,12 @@ export class TasksController {
       return task;
     }
     throw new NotFoundException();
+  }
+
+  private checkTaskOwnership(task: Task, userId: string): void {
+    if (task.userId !== userId) {
+      throw new ForbiddenException('You can only access your own tasks!');
+    }
   }
 }
